rag/requirements.txt
# Core LiveKit dependencies
livekit-agents[openai,silero,deepgram,turn-detector]~=1.0
python-dotenv

# RAG dependencies
llama-index>=0.10.0
llama-index-vector-stores-qdrant>=0.6.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-readers-file>=0.1.0
qdrant-client>=1.14.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.11.0
requests>=2.31.0

# Performance optimizations
asyncio-throttle>=1.0.0